+++ client/src/components/Navbar.js
@@ -12,11 +12,13 @@ export default function Navbar(props) {
   let navigate = useNavigate();
 
   const handlelogout = () => {
-    alert('Are you sure to log out?')
-    localStorage.removeItem("authTkn");
-    localStorage.clear();
-    props.getdata(null)
-    navigate('/')
+
+    if (window.confirm('Are you sure to log out?')) {
+      localStorage.removeItem("authTkn");
+      localStorage.clear();
+      props.getdata(null)
+      navigate('/')
+    }
   }
 
   const role = localStorage.getItem('userrole')
@@ -34,27 +36,27 @@ export default function Navbar(props) {
         <div className="container-fluid">
 
           {
-            
-            (auth!== null)?
-            (role==='admin')?
-            <Link to='/' className='navbar-brand ms-5 text-dark '>eLitmus- Admin</Link>:
-            <Link to = '/' className='navbar-brand ms-5 text-danger'> eLitmus</Link>:
-            <Link to='/' className="navbar-brand ms-5 text-danger">eLitmus</Link>
+
+            (auth !== null) ?
+              (role === 'admin') ?
+                <Link to='/' className='navbar-brand ms-5 text-dark '>eLitmus- Admin</Link> :
+                <Link to='/' className='navbar-brand ms-5 text-danger'> eLitmus</Link> :
+              <Link to='/' className="navbar-brand ms-5 text-danger">eLitmus</Link>
           }
-          
+
 
           <div className="d-flex">
 
             {
-              (localStorage.getItem('authTkn')===null)
+              (localStorage.getItem('authTkn') === null)
                 ?
 
                 <div>
                   <Link to='/signup' className='btn text-dark me-2'>Sign up</Link>
                   <Link to='/login' className='btn btn-success rounded rounded-3 text-white me-5'>Login</Link>
                 </div>
-                
-                :  
+
+                :
 
                 <div>
                   <div className="btn btn-success rounded text-white me-3 bg-dark">Welcome {localStorage.getItem('username')}</div>
